[#283] refactor: badge를 처리하는 주체로 명확한 이름으로 변
